Import linux-style's list API for quick sort implementation
Implemented quick sort using Linux-style List API. The main difference
from the previous implementation in "list.c" is that the Linux-style
List API is designed for circular linked list. As the result, the entire
implementation needed be change to accomodate this. The benefits of the
circular list include saving memory and time when acessing the tail element.
Furthermore, "linx_list_is_ordered()" has been added to check if the list is
sorted as desired.